requirements.txt
alembic==1.2.1
aniso8601==8.0.0
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
paho-mqtt==1.4.0
pycodestyle==2.5.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
redis==3.3.8
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
